+++ CHANGELOG.md
@@ -1,3 +1,29 @@
 ## 0.0.1
-
 * TODO: Describe initial release.
+
+## 0.0.3
+* Fix update devices listener.
+
+## 0.0.4
+* Added abort charge.
+
+## 0.0.5
+* Fix: Reset state before charge.
+
+## 0.0.6
+* Fix: Check distinct stream data.
+
+## 0.0.7
+* Fix: Removed unnecessary stream seed.
+
+## 0.0.8
+* Fix: Clear payment data after transmit it.
+
+## 0.0.9
+* Fix: Revert clear payment data after transmit it.
+
+## 0.0.10
+* Fix: Adjuts in .java.
+
+## 0.0.10
+* Fix: adjust code on success payment.